Fix format strings
Summary:
File tscconfig.cc:

On Windows 64 bits, size_t is wider than unsigned long, therefore
this format string needs to be adapted.  Also make ID generation
thread-safe and prefer `snprintf` over `sprintf` while ensuring
correct string termination.

Other files:
Enhance defense of existing `snprintf` invocations against string
termination errors and future edits.
Correct format string signedness errors.

Test Plan:
Successful test builds on Linux (PC and ARM) and Mac.

Code review.

Reviewers: giso

Reviewed By: giso

Tags: #tascar

Differential Revision: https://dev.openmha.org/D635

+++ gui/src/tascar_mainwindow.cc
@@ -242,8 +242,8 @@ bool tascar_window_t::on_timeout_statusbar()
           TASCAR::scene_render_rt_t* scene(session->scenes[selected_scene]);
           TASCAR::render_profiler_t prof(scene->loadaverage);
           prof.normalize(prof.t_postproc);
-          sprintf(cmp,"scenes: %ld  point sources: %d/%d  diffuse sound fields: %d/%d | jack: %1.1f%% (scene \"%s\" load: %1.1f%% init: %1.1f%%  geo: %1.1f%%  preproc: %1.1f%%  acoustic: %1.1f%%  postproc: %1.1f%%)",
-                  (long int)(session->scenes.size()),
+          sprintf(cmp,"scenes: %zu  point sources: %d/%d  diffuse sound fields: %d/%d | jack: %1.1f%% (scene \"%s\" load: %1.1f%% init: %1.1f%%  geo: %1.1f%%  preproc: %1.1f%%  acoustic: %1.1f%%  postproc: %1.1f%%)",
+                  session->scenes.size(),
                   session->get_active_pointsources(),session->get_total_pointsources(),
                   session->get_active_diffuse_sound_fields(),session->get_total_diffuse_sound_fields(),
                   scene->get_cpu_load(),
